refactor(core): split root attachment modules
